Initialized splash activity
