+++ android/app/src/main/res/layout/activity_main.xml
@@ -7,8 +7,6 @@
     android:background="@color/colorPrimary"
     tools:context="com.chorestory.app.MainActivity">
 
-
-
     <ImageView
         android:id="@+id/logo_image"
         android:layout_width="400dp"
